+++ onlyoffice-wordpress.php
@@ -7,7 +7,7 @@
  * Plugin Name:       ONLYOFFICE
  * Plugin URI:        https://github.com/ONLYOFFICE/onlyoffice-wordpress
  * Description:       Add ONLYOFFICE editor on page
- * Version:           1.0.2
+ * Version:           1.1.0
  * Requires at least: 5.7
  * Requires PHP:      7.4
  * Author:            Ascensio System SIA
@@ -40,7 +40,7 @@
 /**
  * Currently plugin version.
  */
-define( 'ONLYOFFICE_PLUGIN_VERSION', '1.0.2' );
+define( 'ONLYOFFICE_PLUGIN_VERSION', '1.1.0' );
 define( 'ONLYOFFICE_PLUGIN_URL', plugin_dir_url( __FILE__ ) );