+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 
 ## Added
+- support for the connector to work in conjunction with the plugin "Force Lowercase URLs" (https://wordpress.org/plugins/force-lowercase-urls)
 - compatible with classic editor TinyMCE
 
 ## 1.0.2