+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## Added
 - support for the connector to work in conjunction with the plugin "Force Lowercase URLs" (https://wordpress.org/plugins/force-lowercase-urls)
+- setting authorization header
 - compatible with classic editor TinyMCE
 
 ## 1.0.2