+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - support for the connector to work in conjunction with the plugin "Force Lowercase URLs" (https://wordpress.org/plugins/force-lowercase-urls)
 - setting authorization header
 - compatible with classic editor TinyMCE
+- Link to docs cloud
 
 ## 1.0.2
 ## Changed